Rename client-profile -> profile
